todo_backend/serializers.py
from rest_framework import serializers

from todo_backend.models import Task


class TaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = ['id', 'title', 'description', 'done']
